refactor(DestinationLister,TreeDiagram): Move from ui/ to tools/
