+++ src/main/java/net/boyechko/pdf/autoa11y/ui/cli/PdfAutoA11yCLI.java
@@ -36,13 +36,13 @@ import net.boyechko.pdf.autoa11y.core.ProcessingListener;
 import net.boyechko.pdf.autoa11y.core.ProcessingResult;
 import net.boyechko.pdf.autoa11y.core.ProcessingService;
 import net.boyechko.pdf.autoa11y.document.PdfCustodian;
+import net.boyechko.pdf.autoa11y.tools.DestinationLister;
+import net.boyechko.pdf.autoa11y.tools.TreeDiagram;
 import net.boyechko.pdf.autoa11y.ui.AccessibilityReport;
-import net.boyechko.pdf.autoa11y.ui.DestinationLister;
 import net.boyechko.pdf.autoa11y.ui.FormattedListener;
 import net.boyechko.pdf.autoa11y.ui.LoggingListener;
 import net.boyechko.pdf.autoa11y.ui.SidecarConfig;
 import net.boyechko.pdf.autoa11y.ui.StructTreeTablePrinter;
-import net.boyechko.pdf.autoa11y.ui.TreeDiagram;
 import net.boyechko.pdf.autoa11y.ui.VerbosityLevel;
 import net.boyechko.pdf.autoa11y.validation.Check;
 import org.slf4j.Logger;
