+++ src/test/java/net/boyechko/pdf/autoa11y/GoalDrivenIntegrationTest.java
@@ -38,7 +38,7 @@ import net.boyechko.pdf.autoa11y.core.ProcessingService;
 import net.boyechko.pdf.autoa11y.document.Content;
 import net.boyechko.pdf.autoa11y.document.PdfCustodian;
 import net.boyechko.pdf.autoa11y.document.StructTree;
-import net.boyechko.pdf.autoa11y.ui.TreeDiagram;
+import net.boyechko.pdf.autoa11y.tools.TreeDiagram;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
